separate hook detection into multiple files
