+++ Makefile
@@ -1,5 +1,8 @@
 ifneq ($(KERNELRELEASE),)
-	obj-m := detectiontool.o
+	obj-m := mytool.o
+	mytool-objs := detectiontool.o hook_syscalls.o hook_fops.o hook_networks.o
+	HEADERS := $(PWD)/include
+	ccflags-y += -I$(HEADERS)
 
 else
 	KERNELDIR ?=/lib/modules/$(shell uname -r)/build
@@ -14,10 +17,11 @@ all: kernel client
 kernel:
 	$(MAKE) -C $(KERNELDIR) M=$(PWD) modules
 
-client: detectiontool.ko
+client: mytool.ko
 	gcc -fPIC client.c $$(python$(python_version_major).$(python_version_minor)-config --cflags) $$(python$(python_version_major).$(python_version_minor)-config --ldflags) -o client
 endif
 
 clean:
 	$(MAKE) -C $(KERNELDIR) M=$(PWD) clean
 	rm -rf *.ko *.mod.* *.o *.order *.symvers client 
+
