+++ detectiontool.c
@@ -1,20 +1,7 @@
-#include <linux/module.h>
-#include <linux/kernel.h>
-#include <linux/types.h>
-#include <linux/slab.h>
-#include <linux/string.h>
-#include <linux/fs.h>
-#include <linux/sched.h>
-#include <linux/list.h>
-#include <linux/uaccess.h>
-#include <linux/cred.h>
-#include <linux/version.h>
-#include <linux/syscalls.h>
-#include <asm/asm-offsets.h> /* NR_syscalls */
-#include <linux/proc_fs.h>
-#include <linux/net.h>
-#include <net/tcp.h>
-#include <net/udp.h>
+#include "detectiontool.h"
+#include "hook_syscalls.h"
+#include "hook_fops.h"
+#include "hook_networks.h"
 
 MODULE_LICENSE("GPL");
 MODULE_AUTHOR("Joshua Lim, Fai Yew");
@@ -31,8 +18,8 @@ MODULE_AUTHOR("Joshua Lim, Fai Yew");
 #define DETECTMODULES_CMD "detectmods"
 
 // declarations
-static unsigned long *addr_syscall_table;
-static int (*core_kern_text)(unsigned long addr);
+unsigned long *addr_syscall_table;
+int (*core_kern_text)(unsigned long addr);
 
 // prototypes
 static int tool_procfs_entry_init(void);
@@ -135,160 +122,10 @@ static ssize_t tool_procfs_read(struct file *fp,
 // detect hooks
 static void hook_detection(void)
 {
-	unsigned int sys_num = 0;
-	unsigned long addr;
-	unsigned char test[12];
-	struct file *fp;
-	int no_of_syscall_hooks = 0;
-	int no_of_fops_hooks = 0;
-	int no_of_net_hooks = 0;
-	int hooked_functions = 0;
-	char path[6];
-	struct tcp_seq_afinfo *tcp_afinfo;
-	struct udp_seq_afinfo *udp_afinfo;
-
-	// scan sys_call_table
-	printk(KERN_INFO "detection tool: Scanning sys_call_table...\n");
-	for (sys_num = 0; sys_num < NR_syscalls; sys_num++)
-	{
-		addr = addr_syscall_table[sys_num];
-		if (!core_kern_text(addr))
-		{
-			printk(KERN_ALERT "detection tool: Hook detected! (syscall %d)\n", sys_num);
-			no_of_syscall_hooks++;
-		}
-		// detect inline hook with mov and jmp
-		else
-		{
-			memcpy(test, (void *)addr, 12);
-			if (test[0] == 0x48 && test[1] == 0xb8 && test[10] == 0xff && test[11] == 0xe0)
-			{
-				printk(KERN_ALERT "detection tool: Hook detected! (syscall %d)\n", sys_num);
-				no_of_syscall_hooks++;
-			}
-		}
-	}
-	if (no_of_syscall_hooks)
-		printk(KERN_ALERT "detection tool: %d hooked system calls detected!\n", no_of_syscall_hooks);
-	else
-		printk(KERN_INFO "detection tool: No hooked system calls detected.\n");
-
-	// scan fops
-	printk(KERN_INFO "detection tool: Scanning fops of /, /proc and /sys...\n");
-
-#if LINUX_VERSION_CODE >= KERNEL_VERSION(4, 7, 0)
-#define SCAN_FOPS(NAME)                                                                 \
-	sprintf(path, "/%s", #NAME);                                                        \
-	fp = filp_open(path, O_RDONLY, S_IRUSR);                                            \
-	if (IS_ERR(fp))                                                                     \
-		printk(KERN_ERR "detection tool: Failed to open %s!\n", path);                  \
-	if (IS_ERR(fp->f_op))                                                               \
-		printk(KERN_WARNING "detection tool: %s has no fops!\n", path);                 \
-                                                                                        \
-	addr = (unsigned long)fp->f_op->iterate_shared;                                     \
-	if (!core_kern_text(addr))                                                          \
-	{                                                                                   \
-		printk(KERN_ALERT "detection tool: %s fops hook detected!\n", path);            \
-		no_of_fops_hooks++;                                                             \
-	}                                                                                   \
-	else                                                                                \
-	{                                                                                   \
-		memcpy(test, (void *)addr, 12);                                                 \
-		if (test[0] == 0x48 && test[1] == 0xb8 && test[10] == 0xff && test[11] == 0xe0) \
-		{                                                                               \
-			printk(KERN_ALERT "detection tool: %s fops hook detected!\n", path);        \
-			no_of_fops_hooks++;                                                         \
-		}                                                                               \
-	}                                                                                   \
-	filp_close(fp, 0);
-#else
-#define SCAN_FOPS(NAME)                                                                 \
-	sprintf(path, "/%s", #NAME);                                                        \
-	fp = filp_open(path, O_RDONLY, S_IRUSR);                                            \
-	if (IS_ERR(fp))                                                                     \
-		printk(KERN_ERR "detection tool: Failed to open %s!\n", path);                  \
-	if (IS_ERR(fp->f_op))                                                               \
-		printk(KERN_WARNING "detection tool: %s has no fops!\n", path);                 \
-                                                                                        \
-	addr = (unsigned long)fp->f_op->iterate;                                            \
-	if (!core_kern_text(addr))                                                          \
-	{                                                                                   \
-		printk(KERN_ALERT "detection tool: %s fops hook detected!\n", path);            \
-		no_of_fops_hooks++;                                                             \
-	}                                                                                   \
-	else                                                                                \
-	{                                                                                   \
-		memcpy(test, (void *)addr, 12);                                                 \
-		if (test[0] == 0x48 && test[1] == 0xb8 && test[10] == 0xff && test[11] == 0xe0) \
-		{                                                                               \
-			printk(KERN_ALERT "detection tool: %s fops hook detected!\n", path);        \
-			no_of_fops_hooks++;                                                         \
-		}                                                                               \
-	}                                                                                   \
-	filp_close(fp, 0);
-#endif
-
-	SCAN_FOPS()
-	SCAN_FOPS(proc)
-	SCAN_FOPS(sys)
-
-	if (no_of_fops_hooks)
-		printk(KERN_ALERT "detection tool: %d hooked fops detected!\n", no_of_fops_hooks);
-	else
-		printk(KERN_INFO "detection tool: No hooked fops detected.\n");
-
-	// scan network hooks
-	printk(KERN_INFO "detection tool: Scanning for network function hooks...\n");
-
-#define SCAN_PROC_NET(NAME)                                                             \
-	sprintf(path, "/proc/net/%s", #NAME);                                               \
-	fp = filp_open(path, O_RDONLY, 0);                                                  \
-	if (IS_ERR(fp))                                                                     \
-		printk(KERN_ERR "detection tool: Failed to open %s!\n", path);                  \
-	if (IS_ERR(fp->f_path.dentry->d_inode))                                             \
-		printk(KERN_WARNING "detection tool: %s has no afinfo!\n", path);               \
-                                                                                        \
-	/* tcp */                                                                           \
-	if (!strncmp("tcp", #NAME, 3))                                                      \
-	{                                                                                   \
-		tcp_afinfo = PDE_DATA(fp->f_path.dentry->d_inode);                               \
-		addr = (unsigned long)tcp_afinfo->seq_ops.show;                                  \
-	}                                                                                   \
-	/* udp */                                                                           \
-	else                                                                                \
-	{                                                                                   \
-		udp_afinfo = PDE_DATA(fp->f_path.dentry->d_inode);                               \
-		addr = (unsigned long)udp_afinfo->seq_ops.show;                                  \
-	}                                                                                   \
-                                                                                        \
-	if (!core_kern_text(addr))                                                          \
-	{                                                                                   \
-		printk(KERN_ALERT "detection tool: %s function hook detected!\n", path);        \
-		no_of_net_hooks++;                                                              \
-	}                                                                                   \
-	else                                                                                \
-	{                                                                                   \
-		memcpy(test, (void *)addr, 12);                                                 \
-		if (test[0] == 0x48 && test[1] == 0xb8 && test[10] == 0xff && test[11] == 0xe0) \
-		{                                                                               \
-			printk(KERN_ALERT "detection tool: %s function hook detected!\n", path);    \
-			no_of_net_hooks++;                                                          \
-		}                                                                               \
-	}                                                                                   \
-	filp_close(fp, 0);
-
-	SCAN_PROC_NET(tcp)
-	SCAN_PROC_NET(tcp6)
-	SCAN_PROC_NET(udp)
-	SCAN_PROC_NET(udp6)
-
-	if (no_of_net_hooks)
-		printk(KERN_ALERT "detection tool: %d hooked network functions detected!\n", no_of_net_hooks);
-	else
-		printk(KERN_INFO "detection tool: No hooked network functions detected.\n");
-
+	int hooked_functions = scan_sys_call_table();
+	hooked_functions += scan_fops();
+	hooked_functions += scan_networks();
 	// print summary
-	hooked_functions = no_of_fops_hooks + no_of_syscall_hooks + no_of_net_hooks;
 	if (hooked_functions)
 		printk(KERN_ALERT "detection tool: Scanning complete. A total of %d hooked functions on your system was detected.\n", hooked_functions);
 	else
