+++ include/detectiontool.h
@@ -1,0 +1,22 @@
+#ifndef DETECTIONTOOL_H
+#define DETECTIONTOOL_H
+
+#include <linux/module.h>
+#include <linux/kernel.h>
+#include <linux/types.h>
+#include <linux/slab.h>
+#include <linux/string.h>
+#include <linux/fs.h>
+#include <linux/sched.h>
+#include <linux/list.h>
+#include <linux/uaccess.h>
+#include <linux/cred.h>
+#include <linux/version.h>
+#include <linux/syscalls.h>
+#include <asm/asm-offsets.h> /* NR_syscalls */
+#include <linux/proc_fs.h>
+#include <linux/net.h>
+#include <net/tcp.h>
+#include <net/udp.h>
+
+#endif
