+++ include/hook_fops.h
@@ -1,0 +1,3 @@
+#include "detectiontool.h"
+
+int scan_fops(void);
