+++ include/hook_networks.h
@@ -1,0 +1,3 @@
+#include "detectiontool.h"
+
+int scan_networks(void);
