+++ include/hook_syscalls.h
@@ -1,0 +1,3 @@
+#include "detectiontool.h"
+
+int scan_sys_call_table(void);
